working on API, setup env varibles

+++ conf/killServers.sh
@@ -1,15 +1,19 @@
 #!/bin/bash
 
+if [ -z "$DHT_HOMEDIR" ]; then
+	echo "Env variable \$DHT_HOMEDIR is not set"
+	exit -1
+fi
 
 count=0
 echo "Killing servers..."
 while read line
 do
 	if [ $count == "0" ]; then
-		pid=`ssh -n $line ps aux | grep "java Server" | awk '{print $2}'` # CHANGE TO META
+		pid=`ssh -n $line ps aux | grep "MeatDataServer" | awk '{print $2}'` # CHANGE TO META
 		ssh -n $line kill $pid
 	else
-		pid=`ssh -n $line ps aux | grep "java Server" | awk '{print $2}'`
+		pid=`ssh -n $line ps aux | grep "Server" | awk '{print $2}'`
 		ssh -n $line kill $pid
 	fi
 	(( count++ ))
